struct(lib): restructure the library modules
This commit restructures the library to be more conforming to the
mathematical theory of finite elements. This is an experimental choice
that might not pan out. But hey, this is an experimental library for now
so we'll see.
