+++ src/solvers/mod.rs
